cmd/go: move modload.BinDir to internal/load.defaultBinDir
BinDir depended on a global variable being set in modload.Init. This
function is now removed as part of an error to remove global state.

BinDir could have been changed to read cfg.GOPATH on each call (as
defaultBinDir does now), but BinDir doesn't seem to be in scope for
modload.

This change lets us delete the gopath global variable.

For #40775

Change-Id: I7c5956e92a8e5e9125cdcc5b407bd1f4d35f1b70