runtime: ignore ctxt argument to crosscall2 calls to _cgo_panic
SWIG calls _cgo_panic via a crosscall2 call omitting the fourth (ctxt)
argument. Thus Go sees whatever happens to be in the fourth argument
register as ctxt argument.

Normally this is not an issue, as the context will go unused. However,
in rare cases (such as a memory profile sample) we may call the cgo
traceback function, passing a garbage context, which may cause the
traceback function to crash or do something else wrong.

Work around this by explicitly ignoring ctxt on callbacks to _cgo_panic.
We need to move _cgo_panic from runtime/cgo to runtime so it can be
referenced even when cgo is not used.

Fixes #47995

Change-Id: Ie98bd0974ec1c08692e3e4006bc77f6f74052670