crypto/sha256: avoid escaping on generic architectures
For architectures without a specialized implementation (e.g. arm),
the generic implementation always allocates because it does:

	var block = blockGeneric

which causes the compiler to give up trying to analyze block
even though it is technically only ever one implementation.
Instead of a variable, declare a function that wraps blockGeneric
similar to the arm64 codepath when SHA2 is unavailable.

Credit goes to Cuong Manh Le for more specifically identifying
the problem and Keith Randall for suggesting a concrete solution.

Fixes #48055

Change-Id: I1a6a2e028038e051c83fd72b10a8bf4d210df57d