cmd/compile: delay transformAssign if lhs/rhs have typeparam
This also requires that we sometimes delay Select, if the assignments in
the Comm part of the select have not been transformed. There is a quirk
for transformSelect/tcSelect, which currently just drops the Init lists
of case.Comm clauses as they are transformed. This works fine and is
needed, since there is another quirk, which is that a Init list of a
case.Comm clauses will not be exported correctly, since it is just
written out as statements, and then parsed back as a BLOCK. So, even if
transformSelect is delayed, we need to nil out the init lists.

Fixes #48137

Change-Id: I163aa1f999d1e63616280dca807561b12b2aa779