testing: more informative comment on test race code
A `t.Log` call inside asynchrous code invoked by a test can
cause a data race if the test completes first. The race detector
points at this code, but the comment isn't necessarily that
clear about why this can arise.

Adjust the comment to make it clearer for people that
don't necessarily have familiarity with the `testing` package
code.

Change-Id: I4d7b68d8321934ca42a1bcda03ac10ed739d055e