cmd/compile: support -buildmode=c-shared on linux/mips64{,le}
In the shared mode, GOT must be used to access global variables.
To use GOT, the assembly code needs to be rewritten to use GOT
in the assembler implementation. To simplify the implementation
of assembler and prevent information loss caused by rewritting in
the lower pass. We need to limit rewritting the ssa value that are
related to GOT.

For #43264

Change-Id: Idedaf6c81f085955371320f51bca148ffb42a2d8