cmd/compile: support -buildmode=c-shared on linux/mips64{,le}
The modification of these rules is optimization to load/store global
variables. If there are a sequence of loads/stores nearby a global
variable address, the address can only be loaded from GOT once instead
of every time. A LD instruction is required when load an address. Due
to the reduction of times to load address, the number of LD instructions
generated by assembler will be reduced.

For #43264

Change-Id: Idedaf6c81f085955371320f51bca148ffb42a2d8