cmd/compile: support -buildmode=c-shared on linux/mips64le
In the shared library, the access to global variables must
use GOT and r28(gp) saved the GOT address in PIC code, which
conflicts with the use of GP as the Static Base address of
global variables in static code. Therefore, the operation
based on static base address should be turned off in shared
library mode.

For golang/go#43264

Change-Id: Idedaf6c81f085955371320f51bca148ffb42a2d8