go/types: instantiate methods when instantiating Named types
In the API proposal we decided that instantiation must also instantiate
methods. This CL does that, and eliminates the special handling for lazy
instantiation in lookupMethod.

It is possible that we expand an instance before all method signatures
have been type-checked, so for simplicity we introduce a new flag on
Func, 'isIncompleteMethod', which controls whether we must fully
substitute methods before using them. We could avoid this flag by using
some convention for the structure of an incomplete method (such as the
receiver has no position), but in practice using a flag was cleaner and
didn't increase the size of the Func struct.

Checker.later is adjusted to invoke its func immediately when called
with a nil receiver. This captures a common pattern: when we are
type-checking (check != nil) we must defer until all declarations have
been type checked, but when we are invoked from the type API we can
assume that everything is set up (and therefore later is now).

Updates #47916

Change-Id: I352baa6664cd07f61b06924744382897805f9d29