go/types: record all instances, not just inferred instances
This change modifies the way we record instance information. It changes
the types.Info map to use the instantiated *ast.Ident as its key, and
records information for all instances, not just those that were produced
via function type inference. Accordingly, the types.Info field is
renamed to Instances. The Inferred type is renamed to Instance,
with its Sig field changed to Type.

This was largely motivated by suggestions from mdempsky on the go/types
API proposal (#47916). In our analysis, always using the *ast.Ident as
key and recording all instances makes the API easier to understand and
use.

Instance.TArgs is also renamed to TypeArgs, consistent with other name
changes.

Updates #47916

Change-Id: Ic25ad0cfd65fee6b05e513843c3866ee7a77cfe3