cmd/go/internal/test: clarify `-count` default
Count default is only 1 when the test is not cached.

If the test is cached it will not rerun the test. This causes confusion because the only way to run tests without cache is to use `-count=1`. According to the documentation as is `-count=1` should have no effect. Changing this CLI documentation to clarify the behavior.

ref: https://github.com/golang/go/issues/39056?ts=4#issuecomment-631688012
resolves: https://github.com/golang/go/issues/39056

Change-Id: Ie8bfec32191ef5493f3231ed74540316ca6fc746
GitHub-Last-Rev: 3be413fb11c04bcdad4d38d0c310a38710d93181
GitHub-Pull-Request: golang/go#48388