cmd/compile: optimize integer comparisons with 1/-1 on riscv64
Some integer comparisons with 1 and -1 can be rewritten as comparisons
with 0. For example, x < 1 is equivalent to x <= 0. This is an
advantageous transformation on riscv64 because comparisons with zero
do not require a constant to be loaded into a register.

The new rules trigger approximately 800 times when building cmd and
std.

Change-Id: I2ce9821dd7605a660eb71d76e83a61f9bae1bf25