cmd/compile: don't create or instantiate methods on shape types
We should never use or need methods on non-interface shape types. We do
have corresponding helper functions instantiated with the appropriate
shape types that take the dictionary and the shape-based receiver as the
first two arguments. Each helper function has the same name as what the
corresponding method would be, so it's best not to create the methods (which
would create confusion for import/export).

Keith noticed this issue as part of another CL -- I'm just finishing up
the change.

Change-Id: I401fbdad791bdb5792617449cad68aa8df1d9911